refactor(config): split server config tests for GH727
